Extracted Mesh into external class
